requirements.txt
Flask
requests
python-dotenv
pyyaml
Flask-Cors
waitress
gunicorn
uvicorn
asgiref